Retrieve flags from statically defined cli flags
The urfave/cli c.FlagNames() depends on flags that have
been defined from the CLI itself and not on environment
variables. This can cause errors of flags not being set.

This is problematic when using a setup that depends
on environment variables setup like Docker.

Let's base ourselves on the statically defined flags
instead since those values are always defined.