Add a way to handle non-markdown services
Services have different ways of handling Markdown.
(Eg. Telegram) In those cases, it might just be better to
send the raw HTML body. By chance Telegram supports HTML
parsing.

With that said, a better way to handle this could be to
create a plugin in the "html-to-markdown" library
for handling Telegram cases.